refactor: Move StatusViewData to core.data.model
Makes it available to code in other modules, which will be necessary
when implementing cached notifications.
