+++ app/src/main/java/app/pachli/adapter/FilterableStatusViewHolder.kt
@@ -20,13 +20,13 @@ package app.pachli.adapter
 import android.view.View
 import androidx.core.text.HtmlCompat
 import app.pachli.R
+import app.pachli.core.data.model.IStatusViewData
 import app.pachli.core.data.model.StatusDisplayOptions
 import app.pachli.core.model.FilterAction
 import app.pachli.core.network.model.Filter
 import app.pachli.core.network.model.FilterAction as NetworkFilterAction
 import app.pachli.databinding.ItemStatusWrapperBinding
 import app.pachli.interfaces.StatusActionListener
-import app.pachli.viewdata.IStatusViewData
 
 open class FilterableStatusViewHolder<T : IStatusViewData>(
     private val binding: ItemStatusWrapperBinding,
