+++ app/src/main/java/app/pachli/adapter/StatusDetailedViewHolder.kt
@@ -11,6 +11,7 @@ import app.pachli.R
 import app.pachli.core.common.extensions.hide
 import app.pachli.core.common.extensions.show
 import app.pachli.core.data.model.StatusDisplayOptions
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.core.preferences.CardViewMode
 import app.pachli.core.ui.NoUnderlineURLSpan
 import app.pachli.core.ui.createClickableText
@@ -18,7 +19,6 @@ import app.pachli.databinding.ItemStatusDetailedBinding
 import app.pachli.interfaces.StatusActionListener
 import app.pachli.util.description
 import app.pachli.util.icon
-import app.pachli.viewdata.StatusViewData
 import java.text.DateFormat
 import java.util.Locale
