+++ app/src/main/java/app/pachli/components/notifications/StatusNotificationViewHolder.kt
@@ -32,7 +32,9 @@ import app.pachli.core.activity.emojify
 import app.pachli.core.activity.loadAvatar
 import app.pachli.core.common.string.unicodeWrap
 import app.pachli.core.common.util.AbsoluteTimeFormatter
+import app.pachli.core.common.util.SmartLengthInputFilter
 import app.pachli.core.data.model.StatusDisplayOptions
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.network.model.Emoji
 import app.pachli.core.network.model.Notification
@@ -40,10 +42,8 @@ import app.pachli.core.ui.LinkListener
 import app.pachli.core.ui.setClickableText
 import app.pachli.databinding.ItemStatusNotificationBinding
 import app.pachli.interfaces.StatusActionListener
-import app.pachli.util.SmartLengthInputFilter
 import app.pachli.util.getRelativeTimeSpanString
 import app.pachli.viewdata.NotificationViewData
-import app.pachli.viewdata.StatusViewData
 import at.connyduck.sparkbutton.helpers.Utils
 import com.bumptech.glide.Glide
 import java.util.Date
