+++ app/src/main/java/app/pachli/components/report/adapter/StatusViewHolder.kt
@@ -26,7 +26,9 @@ import app.pachli.core.activity.emojify
 import app.pachli.core.common.extensions.hide
 import app.pachli.core.common.extensions.show
 import app.pachli.core.common.util.AbsoluteTimeFormatter
+import app.pachli.core.common.util.shouldTrimStatus
 import app.pachli.core.data.model.StatusDisplayOptions
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.network.model.Emoji
 import app.pachli.core.network.model.HashTag
@@ -39,9 +41,7 @@ import app.pachli.util.StatusViewHelper
 import app.pachli.util.StatusViewHelper.Companion.COLLAPSE_INPUT_FILTER
 import app.pachli.util.StatusViewHelper.Companion.NO_INPUT_FILTER
 import app.pachli.util.getRelativeTimeSpanString
-import app.pachli.util.shouldTrimStatus
 import app.pachli.viewdata.PollViewData
-import app.pachli.viewdata.StatusViewData
 import java.util.Date
 
 open class StatusViewHolder(
