+++ app/src/main/java/app/pachli/components/report/adapter/StatusesAdapter.kt
@@ -23,8 +23,8 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import app.pachli.components.report.model.StatusViewState
 import app.pachli.core.data.model.StatusDisplayOptions
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.databinding.ItemReportStatusBinding
-import app.pachli.viewdata.StatusViewData
 
 class StatusesAdapter(
     private val statusDisplayOptions: StatusDisplayOptions,
