+++ app/src/main/java/app/pachli/components/search/adapter/SearchStatusesAdapter.kt
@@ -22,9 +22,9 @@ import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import app.pachli.adapter.StatusViewHolder
 import app.pachli.core.data.model.StatusDisplayOptions
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.databinding.ItemStatusBinding
 import app.pachli.interfaces.StatusActionListener
-import app.pachli.viewdata.StatusViewData
 
 class SearchStatusesAdapter(
     private val pachliAccountId: Long,
