+++ app/src/main/java/app/pachli/interfaces/StatusActionListener.kt
@@ -18,10 +18,10 @@
 package app.pachli.interfaces
 
 import android.view.View
+import app.pachli.core.data.model.IStatusViewData
 import app.pachli.core.network.model.Poll
 import app.pachli.core.network.model.Status
 import app.pachli.core.ui.LinkListener
-import app.pachli.viewdata.IStatusViewData
 
 interface StatusActionListener<T : IStatusViewData> : LinkListener {
     fun onReply(pachliAccountId: Long, viewData: T)
