+++ app/src/main/java/app/pachli/util/ListStatusAccessibilityDelegate.kt
@@ -10,10 +10,10 @@ import app.pachli.R
 import app.pachli.adapter.FilterableStatusViewHolder
 import app.pachli.adapter.StatusBaseViewHolder
 import app.pachli.core.activity.openLink
+import app.pachli.core.data.model.IStatusViewData
 import app.pachli.core.network.model.Status.Companion.MAX_MEDIA_ATTACHMENTS
 import app.pachli.core.ui.accessibility.PachliRecyclerViewAccessibilityDelegate
 import app.pachli.interfaces.StatusActionListener
-import app.pachli.viewdata.IStatusViewData
 import app.pachli.viewdata.NotificationViewData
 import kotlin.math.min
