+++ app/src/test/java/app/pachli/components/timeline/CachedTimelineViewModelTestStatusFilterAction.kt
@@ -22,8 +22,8 @@ import app.pachli.ContentFilterV1Test.Companion.mockStatus
 import app.pachli.components.timeline.viewmodel.StatusAction
 import app.pachli.components.timeline.viewmodel.StatusActionSuccess
 import app.pachli.components.timeline.viewmodel.UiError
+import app.pachli.core.data.model.StatusViewData
 import app.pachli.core.database.model.TranslationState
-import app.pachli.viewdata.StatusViewData
 import at.connyduck.calladapter.networkresult.NetworkResult
 import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidTest
